+++ Assets/00_Scripts/Turn System/Turn Cycle/04_AiDecideState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36aee8776e3caa3459fb30884cab1f8f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
