+++ Assets/00_Scripts/Turn System/Turn Cycle/05_BattleLoopState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7afccbe918469da4db698aa2db6b8239
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
