+++ Assets/00_Scripts/Turn System/Turn Cycle/06_BattleEndState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0f636bcb8e95d054183fa21e30394c44
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
